requirements.txt
numpy
requests
sgp4
matplotlib
pyserial
lightgbm
